feat: Add astro config
We're basing this off the https://docs.astro.build/en/guides/markdown-content/#markdown-plugins
section of the astro docs, as well as some other remark information. We
have disabled XSS protection early in the rendering pipeline due to a
later sanitization step. This is mentioned on the remark docs and we have
tested it with some basic XSS examples